major changes to post page meta box areas.
You can create new filters from a drop down menu. The option for that
filter is removed when one is selected, and input boxes are generated
on page load for options that have already been saved. Remove option
button will set the option back to default

Note:
1) On page load, the filter option isn't removed from the dropdown for
items that were already set, and
2) The remove button actually removes that item from the array when
saved, but we want it to save null or "" instead, so we know that it's
_not_ there.