Don't send empty 'status' and 'type' parameters to API
Problem:

When the API is sent an empty status parameter, like "..&status=&", it
returns only Active listings. But with not status parameter it returns
Active and Pending listings.

The plugin was setting empty status parameters on short-code pagintation
links when there was no status specified. For example, this would set an
empty status parameter:

[sr_listings type="Land"]

on the pagination links in the output and provide unexpected listings on
the next page.