Add contact form info to the body of the email message
This fixes cases where the contact form's email "From" header gets
overwritten by email providers that don't allow custom values in the
"From" header.

Situation:
In some cases, like Gmail, an email's "From" header is overwritten to be
the email that was used to authenticate with SMTP. Right now, our
contact form relies on the "From" header being set to the email that was
entered with the form.

In these situations, the "From" email address is completely lost which
makes the contact form email basically useless since you don't know how
submitted the form.

Solution:
This adds all of the contact form information (Name, email, and current
listing) to the end of the email message so that if the "From" header is
overwritten, the information is still available in the body of the
message and can be copied out.