Add admin option to show MLS status text
A new field "statusText" has been added to the MlsInformation in the API
response. This adds an option to the admin panel that, if set to true,
will show the raw MLS statusText field in place of the standard status
field.

For example, this may have the effect in some areas of showing the raw
"Contingent" status on a listing that has a standardized status of
"Active".

Note that the statusText can be shown, but not searched. Any status=
searches still work on the standard mls.status even when this option is
true.