Improve logic for displaying baths in [sr_listings_slider]
The [sr_listings] short-code was recently updated to provide better
logic for determing the best field/data to show for the 'bathrooms'
field. This updates the [sr_listings_slider] to use the same logic.

Basically, the 'bathrooms' field is checked first, and if it's
available (and numeric) that is used for the 'baths' number. If that's
not available, 'bathsFull' is used - which is the current behavior.

---

- [x] Improve display logic for `baths` in [sr_listings_slider]
- [x] Bump minor version