Add option to show Agent/Office above the fold
This is a new IDX compliance setting that can be enabled to show Listing
Agent and Listing Office information above-the-fold on single listing
details pages.

This is a requirement and some MLSs, and this option can be enabled to
meet those requirements. The option is not enabled by default, and
normally the Listing Agent/Office is shown in the information table with
the other data.

- [ ] Add option to admin compliance section
- [ ] Implement markup based on admin setting
  - Do we need to check the other settings too, like "Hide agent/office
    contact information"?