Improve function to generate Listing By markup
This adds a function to the SrUtils module that handles creating the
markup for the sr_agent_office_above_the_fold option.

It has it's own function so that it can handle showing the correct
output in various cases, like when agent or office may be missing, but
the other is available, or when neither of them are available.