Update PHP versions in .travis.yml
PHP 5.2-5.3 are no longer supported on Travis-CI's Trusty builds, which
is the default environment. You can use Precise instead of Trusty to
access those PHP versions, but since they are both old and no longer
supported, it's better if we start running our tests on 5.4 and 5.6.

I looked around most hosting providing, like GoDaddy, and their defaults
are now 5.4 and above, so it's good if we start running tests on these
versions anyway.

This commit changes are test suite to run on PHP 5.4 and PHP 5.6 instead
of PHP 5.3 and PHP 5.4