Add admin option to show MLS trademark symbol
This adds an admin Compliance Settings option that, when enabled, shows
the circle R trademark symbol wherever the text "MLS" is shown. Eg:

Enabled: MLS®
Disabled: MLS

This mostly only needs to be used in Canada where "MLS" is a registered
trademark. For USA MLS companies, this generally does not need to be
enabled, so we're making it an option that is disabled by default.