Fix Google Maps error in client side sr_map_search
**BUGFIX**

This fixes a new error in the client side JavaScript that uses the
Google Maps API. Google recently updated the Maps API so that the
function "Map" cannot be redefined. This was the name we used for our
client-side map function, and the new Google update makes this breaks
and causes maps not be shown.

The fix is to simply rename our Map function to SimplyRETSMap, which is
more unique and shouldn't have any issues with conflicting code.