Bumb version number - v2.3.4
This release contains a fix for a critical bug in the [sr_map_search]
short-code. Google updated their maps API to make "Map" a private name
that cannot be overridden, but our plugin was using that name. The fix
is to change the name of our Map function to SimplyRETSMap.