Bump version number - v2.3.5
This release contains a new admin compliance option to show the circle R
trademark symbol everywhere the text "MLS" is shown. This is useful for
Canadian MLS's where the term "MLS" is trademarked.

In the US, this generally shouldn't have to be enabled so the option is
disabled by default and can be enabled in the plugin's admin settings.