Support multiple 'q' params in sr-search post pages
An sr-search page is loaded whenever a user makes a search via the
search form, or whenever they click the "prev/next" pagination links on
search results.

This adds logic to that custom post type to support the "q" parameter
having multiple values.

There are two cases the note here:

- First, if the `sr_q` parameter is available, we know that a user
  clicked the pagination links. `sr_q` is available as an Array, so we
  transform that into query parameters.
- Second, if `sr_keywords` is set instead, the user is coming from a
  search form submission. `sr_keywords` is always a string (even if it
  has multiple values), so it is first split on the semi-colon
  delimeter, then transformated into query parameters.