Support "status" attribute on [sr_search_form] short-code
This adds support for setting a "status" attribute on the
[sr_search_form] short-code so that a form can be configured for
searching specific statuses instead of the default.

NOTE: If the advanced attribute is set to true on [sr_search_form], the
status attribute already works. This adds functionality for the
non-advanced version of the search form.

Usage: [sr_search_form status="ActiveUnderContract"]

**Tasks**

- [ ] Support "status" attribute on non-advanced [sr_search_form]
- [ ] Retain search form status in next/prev pagination links
- [ ] Retain status filter on page 2+ of a search (similar issues in the
  past, just noting it so it can be tested).
- [ ] Support specifying multiple statuses separated by ";" (NOTE: the
  advanced version of the search form only supports 1 status).
- [ ] Advanced search form functionality is the same