Add support for setting initial points on [sr_map_search]
This adds support for a `pointsQuery` attribute on the [sr_map_search]
short-code. This is useful, for example, when a user wants to use the
[sr_map_search] and show listings from a pre-defined area.

Due to the way WP parses short-code attributes, the user must provide a
query-string like value in the `pointsQuery` attribute on the
short-code. That value is then parsed by the client side code and sent
with the initial request when listings are loaded.