Add option to show agent name with compliance markup
This adds an admin option to the Listing Compliance Settings in the
plugin to show the listing agent name on summary thumbnails in addition
to showing the listing office. This works the same as showing the
listing office name, and can be enabled together or separately depending
on the requirements.

- [x] Update `mkListingSummaryCompliance` to accept agent name
- [x] Create compliance markup based on the two settings - agent & office
- [x] Test in list view, slider, widgets, and map thumbnails
- [ ] Pass setting to sr_int_map_search