Fix uncaught exception in 'q' parameter search
**Issue**
Error: Uncaught Error: Unsupported operand types

Basically, the code was - in some cases, trying to add a string and an
array, which is invalid in PHP.

The issue in the code was that `sr_q` was defaulted to an array -
instead of the resolved value being _casted_ to an array. So, if a value
was provided for `sr_q` it was never casted to an array and the
exception was thrown.

**Solution**
More specifically cast `sr_q` and `sr_keywords` to arrays only after
`get_query_var` has been called - which defaults to a string.

- [x] Fix uncaught exception
- [x] Run some test cases