Retain search form 'agent' parameter during pagination
When this parameter is applied to the short-code it works on the first
page, but was not being re-applied to the search form after paginating
to the second page of results and then making another search.

This makes sure any 'agent' parameter is applied to the search form
short-code during pagination.