Add admin option to enable/disable ListHub test events
This adds an option to the admin panel for the user to control if
analytic events sent to ListHub are in test mode or not. We previously
had a flag that was always set to `true` (and I don't think ListHub
required it initially), but this now gives the user the option to set
that to `false` instead - so that events are sent in "live" mode.

- [ ] Add admin option for ListHub test events
- [ ] Bump version number - v2.4.2