Support null address in sr-single permalinks
Null addresses already actually work, this just improves the formatting
of the address passed through the permalink by removing a preceding
comma if the address.full is null.

A listing may have a null address if the RETS feed has settings like
"Disp_addr" or "IDXAddressInclude" and they are set to false. This
updates the plugin to allow linking to pages even if `.address.full` is
`null`.

- [x] Format address correctly if `address.full` is missing on
  `sr-single` pages.
- [x] Check/test all 3 permalink types

---

One other side note is that we could show "Undisclosed address" instead
of nothing, but showing nothing is aesthetically better.