Set better zoom level on maps when only 1 listing is shown
If only one listing is shown on a search results map, the 'auto' zoom
is enabled and it way too far zoomed in to be able to tell what's
going on.

This updates checks when rendering the map if only 1 unique
latitude/longitude pin is being placed. If so, we set a custom zoom
level of 12 around the pin, instead of using the auto zoom.