Update install from .zip instructions in README
Fixes #81

The install from zip installation instructions were incomplete. The
`vendor` artifacts are not bundled with the zip files when downloaded
from GitHub so the instructions would never work.

This updates the instructions to link to the WordPress plugin page to
download the zip file, instead of the GitHub releases page, so that
the vendor artifacts are included with the zip file.

This also adds instructions to run composer install when downloading
from the GH releases page.