Pass raw address string to infoWindowMarkup
The issue is right now, we're passing the address after it's been made
into HTML with the `srDetailsTable` function. This updates the listing
details page to use the raw "address.full" data in the map marker info
window.

You can see the issue in this picture, that the address has extra
markup (the "Address" at the beginning):

[pic]