Query all basic statuses from widget API requests
This improves the SimplyRETS widgets to query for
`ActiveuUnderContract` listings in addition to Active and Pending.

There has been a few cases recently where, in areas that utilize
ActiveUnderContract, the widgets aren't able to find the listings
because the widgets just use default query paramters.

Now, the widgets specifically set `status` parameters for Active,
Pending, and ActiveUnderContract.

In the future, we can get the available statuses for the current
user/vendor from the DB and query all additional statuses the user can access.