Support cities, counties, and neighborhoods parameters on [sr_search_form]
A follow-up to #113 that makes more parameters available with the same
functionality.

You can now specify parameters on the `[sr_search_form]` short-code
for `cities`, `counties`, and `neighborhoods`. For example:

```
[sr_search_form counties="..." cities="..." neighborhoods=".."]
```

This will persist these values in the search form so that a user can
create "area specific" search forms, restricting what the user can
actually search.