Add support for openhouses search results and details pages
Ref #111

:warning: Work in progress :warning:

This adds support for Open Houses in the SimplyRETS plugin, in a
couple of ways:

- [x] `sr_openhouses` short-code
- [ ] Show open houses on listing details pages

The `[sr_openhouses]` short-code allows you to add an open house
search to your website, just like the property search you get when
using `[sr_listings]`. There are still a few things to do here:

- [ ] Make sure all parmaters are supported
- [ ] Finalize/cleanup markup

Showing open houses on listing details pages will be available as a
way to "call out" listings that have an upcoming open house. For users
with EnterpriseAccess, we can check for open houses when a listing
details page is viewed and show a banner at the top of the
page. We can also have an additional table at the bottom of the page
with the open house details.