Support all API parameters on [sr_map_search] short-code
This updates the `[sr_map_search]` short-code to allow the user to
pass any valid API parameters to the short-code as default search
values. This allows users to do things like this:

```
[sr_search_form agent="1; 2" cities="1; 2" status="1; 2" ...etc]
```

Previously we only supported the `limit` parameter, making this
short-code much less useful for most of the common use-cases like
having a search page for a specific agent or area.

This needs a little bit of testing, but it's fairly straightforward:
you can add any attributes to the `[sr_search_form]` short-code, and
they will be sent to the API. The user CAN override default values by
using the search form. Some parameters can be specified on the
short-code but aren't visible in the search form. Those will always
use the default provided and cannot be overwritten by the user.