Handle "/" characters in listing details page link
This adds handling for "/" characters that appear in within the
address of a link to the litsing details pages. We have to create a
special way of handling these because normal URL encoding of
`/`->`%2F` is not supported, by default, in Apache.

Right now the proposed change converts a `/` to `_`, when creating a
link to a details page, and then converts it back from `_` -> `/` when
using this part of the link to render the page title.

The `_` seems like a fairly uncommon character in address strings, so
I thought it would be a safe bet for now. The other consideration for
using a simple character (and not, say `/`->`<slash>`) is that we want
to the address in the URL bar to still be fairly read-able.

---

Ideally, down the road as we continue improving the plugin, we don't
need to rely on the URL to create the page title within WordPress, and
we can simply used data fetched from the API. But the way things are
set up right now with "dynamic post content", this is how we have to
do it. Although, I do see there are better ways of handling dynamic
posts within WordPress that we should work towards.