Don't show `next` link when limit > number of results
We're working on a fix for an API bug where a `next` link is still
returned on the last page when `count=false`.

This addresses that issue by not showing the next link when limit is
greater the number of results. When the API side is fixed this won't
be necessary, but for now it helps decrease error for end users.